requirements.txt
psutil
py-cpuinfo